python/requirements.txt
aniso8601==9.0.1
blinker==1.9.0
cassandra-driver==3.29.2
click==8.1.8
Flask==3.1.0
Flask-RESTful==0.3.10
geomet==0.2.1.post1
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pytz==2024.2
six==1.17.0
Werkzeug==3.1.3
